requirements.txt
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.8.2
CacheControl==0.12.5
cached-property==1.5.1
cachetools==2.1.0
camelot-py==0.7.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
Django==2.1.1
django-appconf==1.0.1
django-ckeditor==5.4.0
django-cors-headers==2.4.0
django-image-cropping==1.1.0
django-imagekit==4.0.2
django-js-asset==1.0.0
django-modeltranslation==0.13b1
django-uwsgi-cache==1.0.1
djangorestframework==3.8.2
drf-yasg==1.9.1
et-xmlfile==1.0.1
firebase-admin==2.12.0
flex==6.13.2
future==0.16.0
google-api-core==1.3.0
google-auth==1.5.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
google-cloud-storage==1.10.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
inflection==0.3.1
isodate==0.6.0
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10
jsonpointer==1.14
jsonschema==2.6.0
lxml==4.5.0
MarkupSafe==1.1.1
msgpack==0.5.6
numpy==1.18.2
openapi-codec==1.3.2
opencv-python==4.2.0.32
openpyxl==3.0.3
pandas==1.0.3
pdfminer.six==20200124
pilkit==2.0
Pillow==5.2.0
protobuf==3.6.0
psycopg2-binary==2.7.4
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycryptodome==3.9.7
PyJWT==1.6.1
PyPDF2==1.26.0
python-dateutil==2.8.1
pytz==2018.3
PyYAML==3.13
requests==2.18.4
requests-toolbelt==0.9.1
rfc3987==1.3.7
rsa==3.4.2
ruamel.yaml==0.15.45
simplejson==3.15.0
six==1.11.0
sortedcontainers==2.1.0
soupsieve==2.0
strict-rfc3339==0.7
swagger-spec-validator==2.3.1
uritemplate==3.0.0
urllib3==1.22
uWSGI==2.0.18
validate-email==1.3
xlrd==1.2.0
zeep==3.3.1
